pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "im12-dt"
version = "0.1.0"
description = "DT-FLOW"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.25",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "einops>=0.7",
    "rich>=13.0",
    "matplotlib>=3.7",
]


[tool.black]
line-length = 130


[tool.isort]
profile = "black"


[tool.flake8]
max-line-length = 130
extend-ignore = ["E203", "W503"]